change directory's name: entiity -> entity change in Person, ImageOfPerson and Executor: replace id to chat_id
