+++ src/main/java/by/polikarpov/backend/repository/ActivityAreaRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.ActivityArea;
+import by.polikarpov.backend.entity.ActivityArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ActivityAreaRepository extends JpaRepository<ActivityArea, Integer> {
