+++ src/main/java/by/polikarpov/backend/repository/ExecutorRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.Executor;
+import by.polikarpov.backend.entity.Executor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExecutorRepository extends JpaRepository<Executor, Integer> {
