+++ src/main/java/by/polikarpov/backend/repository/ImageOfPersonRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.ImageOfPerson;
+import by.polikarpov.backend.entity.ImageOfPerson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageOfPersonRepository extends JpaRepository<ImageOfPerson, Integer> {
