+++ src/main/java/by/polikarpov/backend/repository/ResumeRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.Resume;
+import by.polikarpov.backend.entity.Resume;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResumeRepository extends JpaRepository<Resume, Integer> {
