+++ src/main/java/by/polikarpov/backend/repository/UserStatusRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.UserStatus;
+import by.polikarpov.backend.entity.UserStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
