+++ src/main/java/by/polikarpov/backend/repository/WorkExperienceRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.WorkExperience;
+import by.polikarpov.backend.entity.WorkExperience;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Integer> {
