+++ src/main/java/by/polikarpov/backend/repository/WorkRepository.java
@@ -1,6 +1,6 @@
 package by.polikarpov.backend.repository;
 
-import by.polikarpov.backend.entiity.Work;
+import by.polikarpov.backend.entity.Work;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WorkRepository extends JpaRepository<Work, Integer> {
